Add logging functionality.
Setting logging configuration for file and console logging.

Currently all configuration is handled in the top-level chyme.__init__
file. Not sure if this is a good place to keep it in terms of it
being an API and logging configuration probably being handled by the
calling code, but it will do for now.

File logging code commented out and set to log to console only for
the time being.