Refactor of the TuflowLoader.
Changed so that the initial pass, finding control files and
loading the data, handles ordering instead of the second pass.

It seems neater to have the recursion occur first and separate
the files into different types that list the order the files
should be read. The second pass can then just process the
contents in order. It also means that contents of some files
can always be read before others which is required in some cases.