Reconfigure TuflowField class setup and handle variables.
Change to configuration of the TuflowField type classes.
Mostly just setting them up properly, but also simplifying the
way that they are created.

Create SEStore class for storing and handling scenario and event
values handed in to the loaded.

Setup variable handling so that variables can be read from the
control files and update placeholders when found.

Run a resolve_variables() function after create_components to
convert all variable placeholders and scenario/event values to the
actual value.