Major restructure of TuflowLoader.read() method.
The read method was broken and not ordering the lines properly
when walking the files tree contents. The change includes a major
refactor of the code and a change to the way 'raw file' data is
loaded and stored. Individual files are no longer stored, it
moves straight to ordering by type (tcf, tgc, etc) instead and
each line is stored alongside the TuflowRawFile information giving
its provenance and other metadata.

Logic is fixed and working well (I think!).