(Mostly) completed refactor of file/variable object structure.
Refactored the "instruction" / "variable" / "files" setup of the TUFLOW
API. Add an iodata.py module containing TuflowPartFiles and
TuflowPartVariables to act as storage for single or multiple
files/variables in a single instruction (after the "=="). I think this
makes a lot more sense and means we should be able to abstract the
complexities of handling single/multiple inputs away from the higher
level API.

Change structure of the iofields.py module quite a bit to remove a lot
of the subclassing and prefer composition over inheritence. The object
heirachies were getting a bit complicated and deep. This update removes
them where possible, such as changing the inheritence from the
TuflowPath class to compose the file variable as a TuflowPath instead.

Complete the setup of using the datafactories.py (probably needs a new
name!) to handle loading subdata. For example, loading the contents
of GIS files, csv files, etc. These are setup in initial command
configuration in the TuflowPartFactory and TuflowPartTypes class as an
additional kwarg.

Some more simplification here as well. It's looking a bit cleaner
now, I think.