Major refactor of TUFLOW package module structure.
Update to module names and refactor of class locations to make
the tuflow package model loaders a lot easier to understand.
I think this should be a bit cleaner and allow a nice separation
between the 'raw' inputs and the nicely standardised and cleaned
up final version of the model used in the software.