Setup Chyme error logging observer and interface for client use.
Piggy backs on the logging framework by created a custom Handler
and filter and setting them up in the config file.

Defines a ChymeProgressLogger observer and a Listener interface
that client code can use to setup a class to get realtime feedback
with progress and message updates.

Utility functions included in the logsettings file for adding
the listener to the observer safely.

We probably need to think about the best way to set this up a
little bit more carefully. In most cases the client code would
set up it's own logging configuration and the logging used in
Chyme would get pulled into that. We may need to add some
utilities to add our handler to their logs and alter some
configuration to make it work the way we want (could be an
option that the client can setup?).