Setup parts.json file to load TuflowPartIO configuration.
Removed the messy setup in the TuflowPartFactory out to a json file
in the tuflow package. New TUFLOW commands can now be added to the
json file along with any validators and data loaders they use.

Keeps the lookups and individual setup of different parts out of
the codebase and makes everything a bit nicer.