Create LoadData class and move placeholder resolver to tuflow_utils.
The data relating to the loading process, such as original path
names, parent files, model_roots, file line numbers, etc, are
useful info, but not always relevant. If a new part is created at
somepoint or loaded independently of a model load this information
will not be available.
Separates this data into a LoadData class, so it can be kept with
the TuflowPartIO classes but isolated and not required by the
interface. The LoadData can be setup with basic values required
for all parts, the additional data are option kwargs.

Resolution of placeholder values - like s1, e2 and <<VAL>> - is
useful to access in a number of places.
Move this behaviour to a tuflow_utils function where it can be
accessed from anywhere. This is also cleans up to TuflowPartIO
classes a bit, removing repeated code and making everything neater.